requirements.txt
aiogram==3.4.1
google-cloud-firestore
google-cloud-core
google-auth
aiohttp